+++ database/migrations/2022_10_31_071236_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('unit');
-            $table->unsignedInteger('price');
+            $table->float('price');
             $table->unsignedInteger('amount');
             $table->timestamps();
         });
